divided code into 3 php files
